pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hal"
version = "0.1.0"
description = "Agent evaluation harness with optional benchmark support"
requires-python = ">=3.11"
dependencies = [
    "datasets",
    "weave==0.51.41",
    "gql<4",
    "huggingface-hub",
    "python-dotenv",
    "setuptools>=48.0.0",
    "inspect_ai @ git+https://github.com/UKGovernmentBEIS/inspect_ai@d3a242f74c8a4f42a86d1219eaca12ab620c3be2",
    "inspect-evals @ git+https://github.com/benediktstroebl/inspect_evals@15a81e27ac7161a39e28f247157e36036a61ef18",
    "tenacity>=8.2.0",
    "docker>=7.1.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
swebench = [
    "swebench @ git+https://github.com/SWE-bench/SWE-bench@006a760a95c9cc11e987884d7e311d74a16db88a",
]
azure = [
    "azure-mgmt-compute>=29.1.0",
    "azure-mgmt-network>=25.1.0",
    "azure-mgmt-resource>=23.0.1",
    "azure-identity>=1.12.0",
    "paramiko==3.5.0",
]
appworld = [
    "appworld @ git+https://github.com/stonybrooknlp/appworld.git",
    "appworld-experiments[simplified,smolagents] @ git+https://github.com/stonybrooknlp/appworld.git@main#egg=appworld-experiments&subdirectory=experiments",
]
taubench = [
    "tau-bench @ git+https://github.com/benediktstroebl/tau-bench@807e348b46a225242d5a045a8cecc690719e4b21",
]
scicode = [
  "scicode @ git+https://github.com/peterkirgis/SciCode"
]
assistantbench = [
    "browsergym==0.13.4"
]
hal-agent-inspect = [
    "smolagents @ git+https://github.com/benediktstroebl/smolagents@c77c99a855b13dbef2af66904448948d01bdc25c",
    "tau-bench @ git+https://github.com/benediktstroebl/tau-bench@807e348b46a225242d5a045a8cecc690719e4b21",
    "mammoth",
    "python_pptx>=1.0.2",
    "pdfminer>=20191125",
    "pdfminer.six>=20240706",
    "puremagic>=1.28",
    "pydub",
    "youtube_transcript_api>=0.6.2",
    "SpeechRecognition",
]
hal-agent = [
    "smolagents @ git+https://github.com/benediktstroebl/smolagents@0c62b18dde5a163370919888290ff4f506dc918c",
    "tau-bench @ git+https://github.com/benediktstroebl/tau-bench@807e348b46a225242d5a045a8cecc690719e4b21",
    "mammoth",
    "python_pptx>=1.0.2",
    "pdfminer>=20191125",
    "pdfminer.six>=20240706",
    "puremagic>=1.28",
    "pydub",
    "youtube_transcript_api>=0.6.2",
    "SpeechRecognition",
]
coreagent = [
    "smolagents @ git+https://github.com/huggingface/smolagents@0c62b18dde5a163370919888290ff4f506dc918c",
    "tau-bench @ git+https://github.com/benediktstroebl/tau-bench@bef42de85cdbfb5490e6f3bd19d8e053df977e08",
    "mammoth",
    "python_pptx>=1.0.2",
    "pdfminer>=20191125",
    "pdfminer.six>=20240706",
    "puremagic>=1.28",
    "pydub",
    "youtube_transcript_api>=0.6.2",
    "SpeechRecognition",
]
corebench = [
    "scipy>=1.15.2"
]
colbench = [
    "sweet_rl @ git+https://github.com/facebookresearch/sweet_rl",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "Pillow>=9.0.0",
    "selenium>=4.0.0",
]
open_deep_research = [
    "anthropic>=0.37.1",
    "audioop-lts<1.0; python_version >= \"3.13\"",
    "beautifulsoup4>=4.12.3",
    "datasets>=2.21.0",
    "google_search_results>=2.4.2",
    "huggingface_hub>=0.23.4",
    "mammoth>=1.8.0",
    "markdownify>=0.13.1",
    "numexpr>=2.10.1",
    "numpy>=2.1.2",
    "openai>=1.52.2",
    "openpyxl",
    "pandas>=2.2.3",
    "pathvalidate>=3.2.1",
    "pdfminer>=20191125",
    "pdfminer.six>=20240706",
    "Pillow>=11.0.0",
    "puremagic>=1.28",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "python_pptx>=1.0.2",
    "requests>=2.32.3",
    "serpapi>=0.1.5",
    "tqdm>=4.66.4",
    "torch>=2.2.2",
    "torchvision>=0.17.2",
    "transformers>=4.46.0",
    "youtube_transcript_api>=0.6.2",
    "chess",
    "sympy",
    "pubchempy",
    "Bio",
    "scikit-learn",
    "scipy",
    "pydub",
    "PyPDF2",
    "python-pptx",
    "torch",
    "xlrd",
    "SpeechRecognition",
    "smolagents[quality,test] @ git+https://github.com/huggingface/smolagents@d8daf8f6e68d4e7ebe9f428938c3466fe482ebb0",
    "sqlalchemy", # for ./examples",
]

[project.scripts]
hal-eval = "hal.cli:main"
hal-upload = "hal.utils.upload:upload_results"
hal-decrypt = "hal.utils.decrypt:decrypt_cli"

[tool.setuptools]
packages = ["hal"]
